# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  memoryrouter.MemoriesListResponse:
    properties:
      memories:
        items:
          $ref: '#/definitions/memoryrouter.MemoryListItem'
        type: array
      page:
        $ref: '#/definitions/router.PageInfoDTO'
    type: object
  memoryrouter.MemoryDTO:
    properties:
      default_key_template:
        type: string
      description:
        type: string
      flushing:
        additionalProperties: {}
        type: object
      id:
        type: string
      locking:
        additionalProperties: {}
        type: object
      max_context_ratio:
        type: number
      max_messages:
        type: integer
      max_tokens:
        type: integer
      persistence:
        additionalProperties: {}
        type: object
      privacy_policy:
        additionalProperties: {}
        type: object
      resource:
        type: string
      token_allocation:
        additionalProperties: {}
        type: object
      token_provider:
        additionalProperties: {}
        type: object
      type:
        type: string
    type: object
  memoryrouter.MemoryListItem:
    properties:
      default_key_template:
        type: string
      description:
        type: string
      etag:
        example: abc123
        type: string
      flushing:
        additionalProperties: {}
        type: object
      id:
        type: string
      locking:
        additionalProperties: {}
        type: object
      max_context_ratio:
        type: number
      max_messages:
        type: integer
      max_tokens:
        type: integer
      persistence:
        additionalProperties: {}
        type: object
      privacy_policy:
        additionalProperties: {}
        type: object
      resource:
        type: string
      token_allocation:
        additionalProperties: {}
        type: object
      token_provider:
        additionalProperties: {}
        type: object
      type:
        type: string
    type: object
...
  router.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  router.PageInfoDTO:
    properties:
      limit:
        example: 50
        type: integer
      next_cursor:
        example: v2:after:tool-001
        type: string
      prev_cursor:
        example: v2:before:tool-000
        type: string
      total:
        example: 2
        type: integer
    type: object
  router.ProblemDocument:
    properties:
      code:
        example: invalid_cursor
        type: string
      detail:
        example: Invalid cursor parameter
        type: string
      instance:
        example: /api/v0/workflows
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  router.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/router.ErrorInfo'
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    email: support@compozy.com
    name: Compozy Support
    url: https://github.com/compozy/compozy
  description: Compozy is a Next-level Agentic Orchestration Platform, tasks, and tools. Authentication
    is configurable via the SERVER_AUTH_ENABLED environment variable. When enabled, all endpoints require
    Bearer token authentication. When disabled (typical for development), endpoints are accessible without
    authentication.
  license:
    name: BSL-1.1
    url: https://github.com/compozy/compozy/blob/main/LICENSE
  termsOfService: https://github.com/compozy/compozy
  title: Compozy API
  version: "1.0"
paths:
...
  /memories:
    get:
      consumes:
        - application/json
      description: List memory configurations with cursor pagination.
      parameters:
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Page size (max 500)
          example: 50
          in: query
          name: limit
          type: integer
        - description: Opaque pagination cursor
          in: query
          name: cursor
          type: string
        - description: Filter by memory ID prefix
          in: query
          name: q
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Memories retrieved
          headers:
            Link:
              description: RFC 8288 pagination links for next/prev
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/memoryrouter.MemoriesListResponse'
                type: object
        "400":
          description: Invalid cursor
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: List memories
      tags:
        - memories
  /memories/{memory_id}:
    delete:
      description: Delete a memory configuration. Returns conflict when referenced.
      parameters:
        - description: Memory ID
          example: '"conversation"'
          in: path
          name: memory_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: No Content
          headers:
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            type: string
        "404":
          description: Memory not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Memory referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Delete memory
      tags:
        - memories
    get:
      consumes:
        - application/json
      description: Retrieve a memory configuration by ID.
      parameters:
        - description: Memory ID
          example: '"conversation"'
          in: path
          name: memory_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Memory retrieved
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/memoryrouter.MemoryDTO'
                type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Memory not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Get memory
      tags:
        - memories
    put:
      consumes:
        - application/json
      description: Create a memory configuration when absent or update an existing one using strong ETag
        concurrency.
      parameters:
        - description: Memory ID
          example: '"conversation"'
          in: path
          name: memory_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Strong ETag for optimistic concurrency
          example: '"\"abc123\""'
          in: header
          name: If-Match
          type: string
        - description: Memory configuration payload
          in: body
          name: payload
          required: true
          schema:
            additionalProperties: true
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: Memory updated
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/memoryrouter.MemoryDTO'
                type: object
        "201":
          description: Memory created
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            Location:
              description: Absolute URL for the memory
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/memoryrouter.MemoryDTO'
                type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Memory referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "412":
          description: ETag mismatch
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Create or update memory
      tags:
        - memories
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication (Optional - controlled by SERVER_AUTH_ENABLED environment
      variable. When enabled, all API endpoints require valid authentication. In development/testing,
      authentication may be disabled.)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...